requirements.txt
Django==1.6.1
Fabric==1.8.1
South==0.8.4
Unidecode==0.04.14
django-extensions==1.2.5
django-mptt==0.6.0
django-mptt-admin==0.1.6
ecdsa==0.10
ipython==1.1.0
paramiko==1.12.0
pil==1.1.7
psycopg2==2.5.1
pycrypto==2.6.1
pyreadline==2.0
six==1.4.1
